+++ src/main/java/com/better/CommuteMate/domain/user/repository/UserRepository.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
-    Optional<User> findByUserId(Integer userId);
+    Optional<User> findByUserId(Long userId);
     boolean existsByEmail(String email);
 }
